Creation, deletion, null fields.

+++ app/webapp/src/main/java/cornerstone/webapp/services/account/management/AccountRole.java
@@ -15,7 +15,7 @@ public enum AccountRole {
         return id;
     }
 
-    public void setId(final short id) {
+    public void setId(final int id) {
         this.id = id;
     }
 }
